add url actions: sign/resign
https://git.digtlab.ru/trusted.ru/cryptoarmgost/issues/482

+++ app/components/SignatureAndEncrypt/SignatureAndEncryptRightColumn.tsx
@@ -19,11 +19,14 @@ import { multiDirectOperation, multiReverseOperation } from "../../AC/multiOpera
 import {
   activeSetting, changeDefaultSettings, deleteSetting, saveSettings,
 } from "../../AC/settingsActions";
+import { removeUrlAction } from "../../AC/urlActions";
 import {
-  ARCHIVE, DECRYPT, DSS_ACTIONS, ENCRYPT, GOST_28147, GOST_R3412_2015_K, GOST_R3412_2015_M,
-  HOME_DIR, LOCATION_CERTIFICATE_SELECTION_FOR_ENCRYPT,
-  LOCATION_CERTIFICATE_SELECTION_FOR_SIGNATURE,
-  LOCATION_SETTINGS_CONFIG, LOCATION_SETTINGS_SELECT, REMOVE, SIGN, UNSIGN, UNZIPPING, USER_NAME, VERIFY, SIGNING_OPERATION, ARCHIVATION_OPERATION, ENCRYPTION_OPERATION, MULTI_REVERSE,
+  ARCHIVATION_OPERATION, ARCHIVE, DECRYPT, DSS_ACTIONS, ENCRYPT, ENCRYPTION_OPERATION, GOST_28147,
+  GOST_R3412_2015_K, GOST_R3412_2015_M,
+  HOME_DIR,
+  LOCATION_CERTIFICATE_SELECTION_FOR_ENCRYPT, LOCATION_CERTIFICATE_SELECTION_FOR_SIGNATURE,
+  LOCATION_SETTINGS_CONFIG, LOCATION_SETTINGS_SELECT, MULTI_REVERSE, REMOVE, SIGN,
+  SIGNING_OPERATION, UNSIGN, UNZIPPING, USER_NAME, VERIFY,
 } from "../../constants";
 import { DEFAULT_ID, ISignParams } from "../../reducer/settings";
 import { activeFilesSelector, connectedSelector, filesInTransactionsSelector, loadingRemoteFilesSelector } from "../../selectors";
@@ -425,7 +428,10 @@ class SignatureAndEncryptRightColumnSettings extends React.Component<ISignatureA
                     <div className="col s12 svg_icon_text">{localize("Documents.docmenu_sign", locale)}</div>
                   </div>
 
-                  <div className="col s4 waves-effect waves-cryptoarm" onClick={this.props.removeAllFiles}>
+                  <div className="col s4 waves-effect waves-cryptoarm" onClick={() => {
+                    this.props.removeAllFiles();
+                    removeUrlAction();
+                  }}>
                     <div className="col s12 svg_icon">
                       <a data-position="bottom">
                         <i className="material-icons docmenu cancel" />
@@ -448,7 +454,10 @@ class SignatureAndEncryptRightColumnSettings extends React.Component<ISignatureA
                     <div className="col s12 svg_icon_text">{"Проверить"}</div>
                   </div>
 
-                  <div className="col s4 waves-effect waves-cryptoarm" onClick={this.props.removeAllFiles}>
+                  <div className="col s4 waves-effect waves-cryptoarm" onClick={() => {
+                    this.props.removeAllFiles();
+                    removeUrlAction();
+                  }}>
                     <div className="col s12 svg_icon">
                       <a data-position="bottom">
                         <i className="material-icons docmenu cancel" />
@@ -1210,12 +1219,81 @@ class SignatureAndEncryptRightColumnSettings extends React.Component<ISignatureA
           const newPath = trustedSign.resignFile(file.fullpath, cert, policies, signParams, format, folderOut);
 
           if (newPath) {
-            deleteFile(file.id);
-            selectFile(newPath);
+            if (file.remoteId) {
+
+              if (uploader) {
+                let cms = trustedSign.loadSign(newPath);
+
+                if (cms.isDetached()) {
+                  if (!(cms = trustedSign.setDetachedContent(cms, newPath))) {
+                    throw new Error(("err"));
+                  }
+                }
+
+                const signatureInfo = trustedSign.getSignPropertys(cms);
+
+                const normalyzeSignatureInfo: any[] = [];
+
+                signatureInfo.forEach((info) => {
+                  const subjectCert = info.certs[info.certs.length - 1];
+                  let x509;
+
+                  if (subjectCert.object) {
+                    try {
+                      let cmsContext = subjectCert.object.export(trusted.DataFormat.PEM).toString();
+
+                      cmsContext = cmsContext.replace("-----BEGIN CERTIFICATE-----", "");
+                      cmsContext = cmsContext.replace("-----END CERTIFICATE-----", "");
+                      cmsContext = cmsContext.replace(/\r\n|\n|\r/gm, "");
+
+                      x509 = cmsContext;
+                    } catch (e) {
+                      //
+                    }
+                  }
+
+                  normalyzeSignatureInfo.push({
+                    serialNumber: subjectCert.serial,
+                    subjectFriendlyName: info.subject,
+                    issuerFriendlyName: subjectCert.issuerFriendlyName,
+                    notBefore: new Date(subjectCert.notBefore).getTime(),
+                    notAfter: new Date(subjectCert.notAfter).getTime(),
+                    digestAlgorithm: subjectCert.signatureDigestAlgorithm,
+                    organizationName: subjectCert.organizationName,
+                    signingTime: info.signingTime ? new Date(info.signingTime).getTime() : undefined,
+                    subjectName: subjectCert.subjectName,
+                    issuerName: subjectCert.issuerName,
+                    x509,
+                  });
+                });
+
+                window.request.post({
+                  formData: {
+                    extra: JSON.stringify(file.extra),
+                    file: fs.createReadStream(newPath),
+                    id: file.remoteId,
+                    signers: JSON.stringify(normalyzeSignatureInfo),
+                  },
+                  url: uploader,
+                }, (err) => {
+                  if (err) {
+                    console.log("err", err);
+                  }
+
+                  deleteFile(file.id);
+                },
+                );
+              }
+            } else {
+              deleteFile(file.id);
+              selectFile(newPath);
+            }
           } else {
             res = false;
           }
         });
+
+        removeUrlAction();
 
         if (res) {
           $(".toast-files_resigned").remove();
@@ -1717,7 +1795,7 @@ export default connect((state) => {
     users: state.users.entities,
     transactionDSS: mapToArr(state.transactionDSS.entities),
     policyDSS: state.policyDSS.entities,
-    operationIsRemote: state.urlActions.performed,
+    operationIsRemote: state.urlActions.performed || state.urlActions.performing,
   };
 }, {
   activeFile, activeSetting, changeDefaultSettings, createTransactionDSS, dssPerformOperation, dssOperationConfirmation,
